+++ apps/web/src/routes/__root.tsx
@@ -1,12 +1,15 @@
 import { createRootRoute, Outlet } from '@tanstack/react-router'
 import { ThemeProvider } from '@/components/theme-provider'
 import { Toaster } from '@/components/ui/sonner'
+import { ConfirmDialogProvider } from '@/components/ui/confirm-dialog'
 
 export const Route = createRootRoute({
   component: () => (
     <ThemeProvider attribute="class" defaultTheme="system" enableSystem>
-      <Outlet />
-      <Toaster />
+      <ConfirmDialogProvider>
+        <Outlet />
+        <Toaster />
+      </ConfirmDialogProvider>
     </ThemeProvider>
   ),
 })